# Iterant fold left operators need to handle errors thrown in F[_] context