--- a/monix-tail/shared/src/test/scala/monix/tail/IterantFoldWhileLeftSuite.scala
+++ b/monix-tail/shared/src/test/scala/monix/tail/IterantFoldWhileLeftSuite.scala
@@ -162,7 +162,7 @@ object IterantFoldWhileLeftSuite extends BaseTestSuite {
     assertEquals(ref.runTry, Failure(dummy))
     assertEquals(effect, 1)
   }
-  
+
   test("foldWhileLeftEvalL protects against broken seed") { implicit s =>
     var effect = 0
     val dummy = DummyException("dummy")
@@ -304,7 +304,7 @@ object IterantFoldWhileLeftSuite extends BaseTestSuite {
     assertEquals(r, Failure(dummy))
     assertEquals(effect, 1)
   }
-  
+
   test("findL is consistent with List.find") { implicit s =>
     check3 { (list: List[Int], idx: Int, p: Int => Boolean) =>
       val fa = arbitraryListToIterant[Coeval, Int](list, idx, allowErrors = false)
@@ -343,4 +343,31 @@ object IterantFoldWhileLeftSuite extends BaseTestSuite {
     assertEquals(r, Failure(dummy))
     assertEquals(effect, 1)
   }
+
+  test("foldWhileLeft earlyStop gets called for failing `rest` on Next node") { implicit s =>
+    var effect = 0
+
+    def stop(i: Int): Coeval[Unit] = Coeval { effect = i}
+    val dummy = DummyException("dummy")
+    val node3 = Iterant[Coeval].nextS(3, Coeval.raiseError(dummy), stop(3))
+    val node2 = Iterant[Coeval].nextS(2, Coeval(node3), stop(2))
+    val node1 = Iterant[Coeval].nextS(1, Coeval(node2), stop(1))
+
+    assertEquals(node1.foldWhileLeftL(0)((_, _) => Left(0)).runTry, Failure(dummy))
+    assertEquals(effect, 3)
+  }
+
+  test("foldWhileLeftEvalL earlyStop gets called for failing `rest` on Next node") { implicit s =>
+    var effect = 0
+
+    def stop(i: Int): Coeval[Unit] = Coeval { effect = i}
+    val dummy = DummyException("dummy")
+    val node3 = Iterant[Coeval].nextS(3, Coeval.raiseError(dummy), stop(3))
+    val node2 = Iterant[Coeval].nextS(2, Coeval(node3), stop(2))
+    val node1 = Iterant[Coeval].nextS(1, Coeval(node2), stop(1))
+
+    assertEquals(node1.foldWhileLeftEvalL(Coeval(0))((_, _) => Coeval(Left(0))).runTry, Failure(dummy))
+    assertEquals(effect, 3)
+  }
+
 }