Fixes #563