@Avasil 2018-01-26
I did benchmark for```Iterant#toListL``` to satisfy my curiosity:
```scala
Before:
[info] Benchmark                                 Mode  Cnt       Score      Error  Units
[info] IterantFoldLeftLBenchmark.toListL100     thrpt  100  391092.509 ± 3795.786  ops/s
[info] IterantFoldLeftLBenchmark.toListL10000   thrpt  100    4020.913 ±   30.503  ops/s
[info] IterantFoldLeftLBenchmark.toListL100000  thrpt  100     425.652 ±    1.370  ops/s

After:
[info] Benchmark                                 Mode  Cnt       Score      Error  Units
[info] IterantFoldLeftLBenchmark.toListL100     thrpt  100  361662.963 ± 4809.333  ops/s
[info] IterantFoldLeftLBenchmark.toListL10000   thrpt  100    3952.470 ±   48.976  ops/s
[info] IterantFoldLeftLBenchmark.toListL100000  thrpt  100     376.399 ±    3.289  ops/s
```